fix: 将模型主键和外键从 BigInteger 改为 Integer，修复 SQLite 测试自增失败

+++ backend/app/models/camera.py
@@ -1,12 +1,12 @@
-from sqlalchemy import Column, BigInteger, String, Integer, SmallInteger, DateTime, ForeignKey, func
+from sqlalchemy import Column, Integer, String, SmallInteger, DateTime, ForeignKey, func
 from app.database import Base
 
 
 class Camera(Base):
     __tablename__ = "museum_camera"
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
-    room_id = Column(BigInteger, ForeignKey("museum_storage_room.id"), nullable=False, comment="所属库房")
+    id = Column(Integer, primary_key=True, autoincrement=True)
+    room_id = Column(Integer, ForeignKey("museum_storage_room.id"), nullable=False, comment="所属库房")
     name = Column(String(100), nullable=False, comment="摄像头名称")
     rtsp_url = Column(String(500), nullable=False, comment="RTSP拉流地址")
     segment_duration = Column(Integer, default=10800, comment="视频分段时长(秒)")
