+++ backend/app/models/room.py
@@ -1,11 +1,11 @@
-from sqlalchemy import Column, BigInteger, String, Text, SmallInteger, DateTime, func
+from sqlalchemy import Column, Integer, String, Text, SmallInteger, DateTime, func
 from app.database import Base
 
 
 class StorageRoom(Base):
     __tablename__ = "museum_storage_room"
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     name = Column(String(100), nullable=False, comment="库房名称")
     code = Column(String(50), unique=True, comment="库房编号")
     location = Column(String(200), comment="位置")
